Major update for blind test
For easier maintenance we should use blind test instead of image recognition for exact match. Which means the result will not tell pass or fail, but give a quick snapshot of the build.
A new resource file is used for each product, it contains info such as serial number and button positions.